yes, I am breaking everything; no, I am not sorry
a long time ago, on a hard disk far far away...
I wrote a network based DCPU emulator in C,
and so it was for many a year.
Until one day I got sick of looking at the crazy data model,
and all the extra lines to make it work, and thus it was so.
so much and overwhelming, that I tore it to bits, and now it's C++11 ish.
This lonely road being walked is now a little brighter, more object
oriented, and closer to being a program worthy of running.

- Meisaka

+++ hardware/pdi.cpp
@@ -25,7 +25,7 @@ int pdi_addtxword(struct PDI_port *p, isi_time_t *pt, uint16_t word, isi_time_t 
 	return 0;
 }
 
-int pdi_getword(struct PDI_port *p, uint16_t *word)
+int pdi_getword(struct PDI_port *p, uint32_t *word)
 {
 	int dc = PDI_GETC(p->stat);
 	int rc = PDI_GETR(p->stat);
